requirements.txt
django-heroku
python=3.6.7
pip=19.0.3
pandas==1.0.1
numpy=1.17.2
Keras==2.3.1
tensorflow==2.0.0
streamlit==0.57.2
Pillow==7.0.0
lime==0.1.1.37
matplotlib==3.1.3
scikit-image==0.16.2